refactor: fix brax import file mismatch
File name was called brax, but ICLand also imports from the google library brax
